saa716x: Use interrupts for I2C writes instead of polling.
Up to now, the I2C_STATUS register was polled when waiting for a data
byte to be sent out via I2C. This could lead to high CPU load, e.g.
when tuning during bad reception conditions. With this change the write
access to the I2C is using interrupts when waiting for data to be sent.
Now the CPU load in similar conditions is around 0.

Signed-off-by: Andreas Regel <andreas.regel@gmx.de>

[moved from media/common/saa716x/ to media/pci/saa716x/]
Signed-off-by: Soeren Moch <smoch@web.de>